# server/docs/openapi.yml
openapi: 3.1.0
info:
  title: ACM Website API
  description: API for managing ACM website resources including users, events, projects, and more.
  version: 1.0.0
  contact:
    email: acm@example.com
servers:
  - url: http://localhost:5000/api/v1
...
paths:
...
  /users/my:
    get:
      summary: Get current user
      tags:
        - users
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update current user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /projects:
    get:
      summary: List all projects
      tags:
        - projects
      responses:
        '200':
          description: Successful response
          content:
            application/json:    
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Admin Create a new project
      description: |
        Only available for admin users. Creates a new project.
      tags:
        - projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Created
          content:
            application/json:    
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /equipment-rentals:
    get:
      summary: Admin List all equipment rentals
      description: |
        Only available for admin users. Returns a list of all equipment rentals.
      tags:
        - equipment
      responses:
        '200':
          description: Successful response
          content:
            application/json:    
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentRental'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new equipment rental
      tags:
        - equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRental'
      responses:
        '201':
          description: Created
          content:
            application/json:    
              schema:
                $ref: '#/components/schemas/EquipmentRental'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        major:
          type: string
        grad_date:
          type: string
          format: date
        interests:
          type: array
          items:
            type: string
            enum: [web_development, machine_learning, cloud_computing, artificial_intelligence]
        profile_pic:
          type: string
        education_level:
          type: string
          enum: [undergraduate, graduate]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        github_link:
          type: string
    EquipmentRental:
      type: object
      properties:
        item_id:
          type: integer
        user_id:
          type: integer
        date_borrowed:
          type: string
          format: date
        return_date:
          type: string
          format: date
        price:
          type: number
        condition:
          type: string
          enum: [ready, broken, in_maintenance]
...